requirements.txt
Django==1.7.4
MySQL-python==1.2.5
djangorestframework==3.0.4
